Add source code for "elewa" in test.ipynb notebook
This change adds the code "elewa" to the source field in the test.ipynb
notebook. The purpose of this change is to include the specific source
code "elewa" as part of the notebook. This will enable the notebook to
execute the code and produce the desired output. This change has no
significant consequences or considerations. This commit addresses issue
#123.